Final Demo 1 ready backend

+++ Backend/property-management/src/main/kotlin/com/example/propertymanagement/repository/QuoteRepository.kt
@@ -3,4 +3,4 @@ package com.example.propertymanagement.repository
 import com.example.propertymanagement.model.Quote
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface QuoteRepository : JpaRepository<Quote, Long>
+interface QuoteRepository : JpaRepository<Quote, Int>
